Updated To New Syntax & Fixed Undefined Data Error
