+++ antt3.github.io/src/routes/+page.svelte
@@ -1,2 +1,5 @@
-<h1>Welcome to SvelteKit</h1>
-<p>Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation</p>
+<script>
+    export let data
+</script>
+
+<pre>{JSON.stringify(data.projects, null, 2)}</pre>
