--- antt3.github.io/src/routes/index.svelte
@@ -1,33 +1,0 @@
-<script context="module">
-    import { gql, GraphQLClient } from 'graphql-request'
-
-    export const load = async () => {
-        
-        const client = new GraphQLClient(
-            import.meta.events.VITE_GRAPHQL_API
-        )
-
-        const query = gql`
-            query GetProjects {
-                projects {
-                    name
-                    slug
-                    description
-                    demo
-                    sourceCode
-                    image {
-                        url
-                    }
-                }
-            }
-        `
-
-        const { projects } = await client.request(query)
-
-        return {
-            props: {
-                projects,
-            },
-        }
-    }
-</script>
